Refactored css and js
