+++ assets/css/app/_base.scss
@@ -1,11 +1,9 @@
 @import url('https://fonts.googleapis.com/css2?family=Public+Sans:ital,wght@0,400;0,700;1,400;1,700&display=swap');
 
-
 body {
   font-family: 'Public Sans', sans-serif;
   background-color: #dedede;
 }
-
 
 .required {
   color: #ff0000;
