+++ assets/css/app/_progress_bar.scss
