+++ assets/css/app/index.scss
@@ -1,0 +1,4 @@
+@import "../shared";
+
+@import "base";
+@import "progress_bar";
