+++ assets/css/public/_challenge-tabs.scss
