+++ assets/css/public/_details-page.scss
