+++ assets/css/public/_misc.scss
@@ -1,0 +1,85 @@
+@font-face {
+  font-family: "Galano Grotesque";
+  src: url("/fonts/galano_grotesque-webfont.woff2") format("woff2"),
+    url("/fonts/galano_grotesque-webfont.woff") format("woff"),
+    url("/fonts/galano_grotesque-webfont.ttf") format("truetype");
+  font-weight: normal;
+  font-style: normal;
+}
+@font-face {
+  font-family: "Galano Grotesque Light";
+  src: url("/fonts/galano_grotesque_light-webfont.woff2") format("woff2"),
+    url("/fonts/galano_grotesque_light-webfont.woff") format("woff"),
+    url("/fonts/galano_grotesque_light-webfont.ttf") format("truetype");
+  font-weight: normal;
+  font-style: normal;
+}
+@font-face {
+  font-family: "Galano Grotesque Medium";
+  src: url("/fonts/galano_grotesque_medium-webfont.woff2") format("woff2"),
+    url("/fonts/galano_grotesque_medium-webfont.woff") format("woff"),
+    url("/fonts/galano_grotesque_medium-webfont.ttf") format("truetype");
+  font-weight: normal;
+  font-style: normal;
+}
+@font-face {
+  font-family: "Galano Grotesque Semi Bold";
+  src: url("/fonts/galano_grotesque_semi_bold-webfont.woff2") format("woff2"),
+    url("/fonts/galano_grotesque_semi_bold-webfont.woff") format("woff"),
+    url("/fonts/galano_grotesque_semi_bold-webfont.ttf") format("truetype");
+  font-weight: normal;
+  font-style: normal;
+}
+@font-face {
+  font-family: "Galano Grotesque Bold";
+  src: url("/fonts/galano_grotesque_bold-webfont.woff2") format("woff2"),
+    url("/fonts/galano_grotesque_bold-webfont.woff") format("woff"),
+    url("/fonts/galano_grotesque_bold-webfont.ttf") format("truetype");
+  font-weight: normal;
+  font-style: normal;
+}
+@font-face {
+  font-family: "icomoon";
+  src: url("/fonts/icomoon.eot?vfbzl8");
+  src: url("/fonts/icomoon.eot?vfbzl8#iefix") format("embedded-opentype"),
+    url("/fonts/icomoon.ttf?vfbzl8") format("truetype"), url("/fonts/icomoon.woff?vfbzl8") format("woff"),
+    url("/fonts/icomoon.svg?vfbzl8#icomoon") format("svg");
+  font-weight: normal;
+  font-style: normal;
+}
+
+/* Custom Overrides */
+.page-item.active .page-link {
+  /* background: #46b3ba;
+  border-color: #46b3ba; */
+}
+.page-item .page-link {
+  /* color: #46b3ba; */
+}
+.truncate a {
+  /* color: #46b3ba !important; */
+  text-decoration: underline;
+}
+.agency-challenges {
+  margin-top: 30px;
+}
+
+/* templates/session/new.html.eex */
+
+.rules-of-behavior {
+  margin: 60px auto;
+}
+
+.rules-wrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.rules-text {
+  margin: 20px 0;
+}
+
+.center-btn {
+  align-self: center;
+}
