+++ assets/css/shared/_uswds-overrides.scss
@@ -45,3 +45,6 @@
     }
   }
 }
+body {
+  background-color: red;
+}
