+++ assets/css/shared/index.scss
@@ -1,0 +1,1 @@
+@import "uswds-overrides";
