+++ assets/js/app/_challenge_wizard_rules.js
