+++ assets/js/app/_custom_url_generator.js
