+++ assets/js/app/_remove_reports_error.js
