+++ assets/js/app/_rich_text_editor.js
