+++ assets/js/app/index.js
@@ -1,0 +1,12 @@
+import "./_challenge_wizard_rules.js";
+import "./_character_limit.js";
+import "./_custom_select.js";
+import "./_custom_url_generator.js";
+import "./_datetime_field.js";
+import "./_dynamic_nested_fields.js";
+import "./_misc.js";
+import "./_radio_button_show.js";
+import "./_remove_reports_error.js";
+import "./_rich_text_editor.js";
+import "./_section_file_upload.js";
+import "./_solution_file_upload.js";
