+++ assets/js/public/index.js
@@ -1,0 +1,2 @@
+import "./_custom.js"
+import "./_scroll_to_anchor.js"
